--- backend/src/main.ts
@@ -1,9 +1,0 @@
-import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
-
-async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
-  await app.listen(8000);
-}
-
-bootstrap();
