+++ frontend/src/components/controls/csvInput/type.ts
@@ -1,0 +1,3 @@
+export interface CSVInputProps {
+  onValueChange: any;
+}
